Simplify and optimize GatherElements
As explained in the docs for NumPy's equivalent `take_along_axis` operation,
`GatherElements` amounts to iterating over matching 1D slices of the input and
indices, and gathering values from the input using indexes in indices.

For the common case of gathering from the last axis of contiguous inputs, these
1D slices can be obtained by just splitting the inputs into chunks. In other
cases we can use the `lanes` iterator.