Combine profiling summary for main graph and subgraphs
Combine operator timings from the main graph and subgraphs into a single
summary, instead of printing output after each sub-graph run. This makes it
easier to see where time was spent in a model that has many subgraphs.

Instead of printing op timings in `Graph::run_plan`, make that method take a
`Profiler` which collects metrics. The profiler is allocated once at the start
of the main graph run, passed along to sub-graphs when they execute, and a
single summary is printed at the end of the main graph run.