Make `Operator::run_in_place` take an `OpRunContext`
Update the `run_in_place` API to align with recent changes to `Operator::run` by
taking an `&OpRunContext` rather than separate arguments for the pool and
inputs. When additional services or information is exposed to operators via
`OpRunContext` in future, this will make it possible to use that in in-place
execution.