Refactor graph fusions to reduce graph traversals
0296a52b56de0cc5cab9c22a9915c98e21e798c7 refactored unary op fusions to make it
easier to define new fusions and reduce the number of graph traversals. This
commit extends that to more general fusions via a new `FusionVisitor` trait.

The result is that fusions are now more modular and new fusions can be added
without adding an extra pass over the graph for each one. Multiple passes may
are still needed in cases where fusions should be combined.