Split error types for operations and value casts
Split the error types used for when value casts fail (Input -> Tensor) versus
when operations fail. This means that the `TryFrom` impls for `Input`/`Output`
now better represent the failures that are possible, and `OpError` can decorate
cast failures with additional context. When an input cast fails, the input index
is now included in the error.