Account for first input being omitted in `InputList` methods
When an operator is run in-place input indices passed to `InputList` methods are
offset by one because the first input is passed separately.  Account for this
when generating errors so that error messages reflect the real index of the
input rather than the one passed to eg.  `InputList::require`, which will be one
lower.