Remove ability to specify output row stride in GEMM
In all current uses the output row stride was always equal to the output column
count. Allowing a custom row stride complicates handling of uninitialized output
buffers because the implementation would need to ensure that any excess elements
in each output row, if the row stride is larger than the minimum, are
initialized.