Optimize and refactor inner dimension iterators
Re-implement inner dimension iterators to be more efficient and use the same
code to handle both static and dynamic-rank inner views. The new version uses
the `Offsets` iterator to iterate over the start offset for each inner view.
This is more efficient than iterating over outer indices and slicing the tensor
at each step. This is beneficial mainly when the ratio of outer dimension steps
is high relative to the number of inner dimension steps.