Re-implement `Lane` on top of `NdTensorView<T, 1>`
A `Lane` is essentially an iterator over a strided slice, and `NdTensorView` is
a strided slice. Building `Lane` on top simplifies reasoning about safety
because the rules are well defined for views (no view may logically overlap with a
mutable view). A similar change has been made for `LaneMut`.

With this, an obviously redundant bounds check in `Lane::next` can be removed.
This restores performance in `AddSoftmax` after changing it to add potentially
non-contiguous lanes.