Fix combining of overlapping fusions
A subgraph like:

`Div(MatMul(Transpose(X), Transpose(Y)), Scale)`

Can be fused into a single `TransformInputs(FusedMatMul)` node, first by
fusing Div + MatMul, then fusing the transposes.

This was broken however after commit 73b65d9bd024962b09a328ae7768ecc2ba59d4a5
because the optimizer would visit the MatMul first and fuse MatMul +
Transpose, then when visiting the subsequent Div operator, it would try
to fuse the Div + the same MatMul that was already fused.

To make this work the optimizer needs to:

 - Ensure that once a fusion is selected, any of the included operators are not
   included in future fusions during the same traversal.

 - Walk the operator graph in reverse toplogical order. Since fusion visitors
   expect to visit the last/output operator of the fusion, this makes fusions
   work as expected when there is an overlap (eg.  MatMulScaleFusion and
   TransposeFusion may both attempt to fuse the same MatMul), or when a fusion
   has optional output nodes (eg.  LayerNormalizationFusin can match
   with/without a bias).

To get the reverse topological order, an execution plan is created and reversed.
A consequence of this is that any operators in the graph which are unused by
the plan will be ignored and not fused. This is probably a good thing since it
avoids spending cycles fusing operations that will likely not be used ("likely"
because RTen allows callers to specify arbitrary nodes as inputs and outputs,
although the optimizer doesn't promise to preserve them).

Fixes https://github.com/robertknight/rten/issues/767