Fix invalid use of memcpy when cloning tensor with non-Copy element type
Fix a crash when cloning a non-contiguous tensor with a stride-1 inner axis and
non-Copy element type due to invalid use of `std::ptr::copy_nonoverlapping`.
The issue went unnoticed since all usage of tensors in RTen has involved Copy
types until now.

To preserve performance for common element types that are `Copy` (`f32` etc),
the typeid crate is used to specialize on known Copy types, since the compiler
is not able to translate the generic zip+Clone loop into a memcpy.