Fix panic in quantized depthwise convolution reading zero point
The depthwise conv kernel expects the zero point slice passed to it to be the
zero points for the full channel range, not just the range of channels that the
kernel has been asked to compute. The issue only occurs when there are enough
elements in the input to trigger multi-threaded operation.

This fixes a panic when running a quantized version of the models for the Piper
example.