Fix execution planning for subgraphs in optimizer
When generating the reverse-order execution plan in
`GraphOptimizer::apply_fusions`, assume that values captured from parent graphs
will be available.

This fixes an issue in the Whisper example where Transpose -> MatMul fusion was
not applied for cross-attention
(/model/decoder/layers.0/encoder_attn/Transpose_1), resulting in most of the
decoder execution time being spent in that transpose. This is because the
transpose input comes from a captured KV-cache input.

The implementation changes the defaults for `PlanOptions` to assume captures are
available, as that is true during "normal" execution and only false when running
constant propagation, and changes `Graph::execution_plan` to take a
`PlanOptions` which is set to the default in `GraphOptimizer::apply_fusions`.