Enable fused Transpose -> Concat operations to run in place
Efficient KV-cache extension in rten-generate relies on the `Concat` operations
which extend KV-cache inputs running in-place. This broke in the Whisper example
after enabling Transpose -> Concat fusion because the KV-cache concat ops were
fused and did not support running in-place afterwards. As a result a new
KV-cache buffer was getting allocated for each run.

The fix is to support running `TransformInputs` ops in place, as long as
the transforms don't apply to the first/in-place input.