Add MatMulIntegerToFloat fused op
Add an operation that fuses the `Mul(Cast<to=float>(MatMulInteger(a, b, a_zero,
b_zero)), scale)` subgraph produced by applying dynamic quantization to a model.

Compared to the un-fused graph, the benefits of the current implementation are
to fuse the Mul + Cast and to do this in-place.  Further improvements are
possible by pushing the cast down into the matmul kernel.