Move input ID -> operator ID map to `Graph`
Move the hash map used for looking up consumers of a value or constant ID from
`GraphMutator` down into `Graph`. The immediate motivation for this is to enable
fusion visitors to make decisions based on the consumers of an operator. They
currently receive a `Graph` rather than a `GraphMutator`. We could pass them
an `&GraphMutator` instead, but input ID -> operator ID seems generally useful
enough to make it part of the `Graph` struct.

A downside to this change is that it means this hash map is retained after graph
optimization is performed, but isn't currently used afterwards.