Don't store constant node IDs explicitly in resolved value set
The optimizer does a lot of micro-planning where it computes a plan to run a
small subgraph. At the start of planning the planner added the ID of every
constant node in the whole graph to the resolved value set, which is wasteful
when considering a small subgraph. Fix this by creating a custom struct for the
resolved value set, which checks for a node ID being a constant when a lookup is
performed.

This makes loading a quantized whisper-v3-large-turbo model about 20% faster,
with prepacking and mmap-ing enabled.