Add ShapeSliceToConstant fusion
Add a fusion that converts `Slice(Shape(X))` to a constant if the dimensions of
X extracted by the `Slice` are all constant values. This enables constant
propagation to precompute values that depend on the extracted dimension size,
which in turn can enable additional fusions.

The motivating use case is to enable Transpose + FusedMatMul fusion for
cross-attention matmuls in the Whisper encoder.