Prepare Transpose fusion to be generalized
Refactor transpose fusion to make it easier to generalize to support fusing
other layout transformations on inputs and avoid looping over the operator's
inputs twice. The original reason for the first loop was to avoid allocations if
we don't find any input transforms that can be fused, but this is now achieved a
different way.