Fix race condition in `TensorPool::alloc`
Fix a race condition where the buffer list could be modified in between a thread
finding a suitable buffer in the pool and removing it from the pool. This could
result in `alloc` returning a different buffer than intended, which might have
less capacity than requested. The condition did not trigger in practice because
all ops currently allocate on the main thread (where `Graph::run_plan`
executes), but was found while making some changes to the `Split` operator which
did result in allocations happening from different threads.

This race was possible because the mutex was released in-between these steps.
The fix is to avoid releasing it until we fall back to allocating via the global
allocator.