Enable i8mm by default
On Apple Silicon (tested on M3) the difference between i8mm and dotprod is
either zero or very marginal. The kernel itself is very slightly faster, but
this can be outweighed by interactions between the parallelism strategy and tile
shapes. On other CPUs (eg. Graviton) i8mm is significantly faster than dotprod.

For the sake of simplicity, enable i8mm everywhere that it is supported.