Use a shared epilogue for int8 GEMM kernels
The int8 GEMM kernels have a different structure for the loop over K depending
on the instructions they use, but the epilogue for adjusting the accumulators to
reflect the zero points and writing results back to memory can be the same for
each.

There is one functional change which is that the kernels using dot product
instructions now initialize the accumulators to zero and add the `k *
a_zero_point * b_zero_point` term in the epilogue, instead of initializing the
accumulators with this term.