Cache names of values captured by an operator's subgraphs
`Graph::capture_names` was taking up ~1.7% of execution time with a Whisper base
model, called via `Graph::run_plan => Graph::operator_dependencies`. Since an
operator's subgraphs are immutable after an `OperatorNode` is created, we can
cache capture names there. In the process replace some `Vec`s with boxed slices
to keep the struct size down.

An alternative approach would be to cache operator dependencies (as `NodeId`s
directly) for the whole graph, but that will require a bit more care with
invalidation in case a captured value node is replaced.