Split int8 packing into main/tail loops
Refactor packing so there are fewer branches in the main loop over full-size
panels and tiles. There are more unsafe blocks now, but these are also more
narrowly scoped. This is also a step towards vectorizing packing.

On a whisper-v3-large-turbo test, this reduced time spent in int8 LHS packing
from 4.4 to 3.3% of total time.