Change shape type for values from slice to SmallVec
Change the type returned by `<Value as Layout>::shape` and `<ValueView
as Layout>::shape` from `[usize]` to `SmallVec<[usize; 4]>`.

This change will support the addition of new value types (eg. sequences) where
the shape is not conveniently stored in a slice or array that we can get a
reference to.

This does add overhead to each `shape` call, but method is not called in
performance-sensitive contexts.