Truncate slices returned by `TensorBase::data` etc. to layout size
TensorBase has an invariant that the length of the storage is at least the
minimum required by the layout, but it can be larger. If the layout was
contiguous and the storage length was larger than required then the `data`,
`data_mut` and `into_slice_mut` methods would return slices that were too long.

This issue rarely surfaced because most `TensorBase` instances have storage that
is exactly the minimum length. However it is possible to create instances where
this is not the case. One way this is currently possible is by broadcasting a
contiguous tensor to a shape with size zero. This issue manifested in a panic
in an `Expand` operator which attempted to do this.