Support `ceil_mode` attribute for AveragePool, MaxPool ops
For the case of "same" padding there is a quirk that two different but
equivalent formulas are given for the output shape in the spec. The main ML
frameworks on which ONNX is based (TensorFlow and PyTorch) individually support
either "same" padding or a `ceil_mode` argument but not both, so this looks like
a corner case that is unimportant.

 - Add `ceil_mode` attribute to MaxPool, AveragePool
 - Implement ceil mode support in `calc_output_size_and_padding`
 - Refactor `calc_output_size_and_padding` tests to use a Default impl
   to save some repetition