Optimize ReduceMin / ReduceMax for float tensors
Reimplement generic kernels for ReduceMin / ReduceMax to use tree reductions of
array chunks. This avoids the serialization of operations that happens otherwise
because float min/max are not associative.

In a LayoutLMv3 model [^1] with `sequence_length=64` this reduced time in
ReduceMax ops from ~33ms to ~4.5ms (M3 Pro).

As part of this, unify implementations of functions in `slice_reductions.rs` and
take advantage of `slice::as_chunks` that was stabilized in Rust v1.88.

[^1]: https://huggingface.co/microsoft/layoutlmv3-base