Refactor mask types and operations in rten-simd
Change the APIs for working with masks to be more like other SIMD vectors:

 - Add `M32`, `M16`, `M8` associated types to `Isa` trait. These are
   types of masks for lanes with 32, 16 and 8-bit elements respectively.

 - Add `m32`, `m16`, `m8` methods to `Isa` trait for getting operations
   on these mask types.

 - Define the mask type of other `Isa` associated types to be M32, M16 or M8
   according to the SIMD vector element width.

This makes the APIs more coherent and allows relating the masks used for
different SIMD vectors. The compiler can now understand for example that vectors
of f32, i32 and u32 elements all use the same mask type.

In order to implement this for AVX2, the mask types had to be changed so that
all vectors of a given lane width use the same mask type. This aligns the AVX2
implementation with how masks are handled for other ISAs.