Add CLI options to read inputs from Safetensors file and print outputs
Add a `--data` option which specifies the path to a Safetensors file from which
the CLI should read input tensor values and a `--print-outputs` option which
prints the debug representation of output tensors.

Together these options are useful to verify that the outputs of running a model
in RTen matches that produced by another system (PyTorch, ONNX Runtime etc.).
The general workflow is:

 1. Modify reference implementation of model in Python to save model inputs
    to a Safetensors file just before running the model, and print the
    outputs immediately after running.

 2. Run `rten model.rten --data inputs.safetensors --print-outputs` and verify
    that the outputs match the reference implementation.

The `--print-outputs` option only prints an abbreviated tensor representation,
so is only useful as a quick sanity check. An option to save the full outputs
for a more thorough comparison will come later.