Add CLI option to compare outputs with reference values
Add a `--check-outputs` argument to the CLI which can be used to check the
results of RTen's inference against reference outputs. This must be used
together with `--inputs` (renamed from `--data` to be more specific).

This is useful for verifying that RTen is producing the same results as PyTorch,
ORT etc. given the same inputs.