Revise `SimdUnaryOp` to remove the need for `same_cast`s
Add a `GetSimd` trait which is implemented for element types and returns the
`Simd` vector type that contains that element in a given `Isa`. This is similar
to `GetNumOps` etc.

Then use this in the `SimdUnaryOp` definition to allow it to know which of
`Isa`'s associated types will be passed and returned according to the element
type. This simplifies the implementation of unary operations as they no longer
need to use `same_cast`.