Fix operator inputs not being added to captures list
Fix a bug when converting subgraphs where captured values were not added to the
capture list in some cases. All operator inputs which do not appear in a graph's
inputs or operator outputs should be marked as captures. However this did not
happen if the input had an entry in the `value_info` field of the GraphProto.

This is because values were marked as captures only if there was no existing
constant or value node with that name. Entries in `value_info` create value
nodes however. The fix is to instead test if the value is "available" from the
local graph (provided as an input, or is computable from operator outputs).

Fixes https://github.com/robertknight/rten/issues/900