Add a minimum size threshold for using memcpy to copy inner lane of tensors
For operators such as `Split`, using memcpy to copy chunks is worthwhile only if
the chunk size is large enough, otherwise the overhead of the function call
makes this slower than using the general nested loop.

If for example we have a tensor like `[X, Y, Z, N]` and are creating N splits
using the last axis, it is better to use the nested loop. Adding a threshold
reduced time in Split operator in NeuFlow v2 from ~21ms to ~4ms.

The threshold here has been picked on a "seems reasonable" basis. Tuning it more
carefully is left as a future exercise.