Support conversion of ONNX attributes with "floats" type
Conversion of attribute values with type
`onnx.AttributeProto.AttributeType.FLOATS` failed because a mapping was missing
in the `_value_fields` dict.

We don't need to implement this mapping ourselves, because there is an
`onnx.helper.get_attribute_value` function, so use that instead. There is also a
`get_node_attr_value` helper, but that doesn't let us verify that the attribute
has the expected type. That helper does however check that there is exactly one
attribute of the given name. For consistency, add the same check in
`AttributeReader.get_attr`.

This fixes an error when converting a ConvNeXt model exported with:

```
./tools/export-timm-model.py timm/convnext_base.fb_in1k --dynamo
```