Handle empty operator output names in model converter
The converter handled empty input names when converting operators, which
represent omitted optional outputs, but not empty output names, which represent
unused outputs. The logic in the converter for serializing the FlatBuffers model
already handled these, as does the model-loading code in Rust. Only the
`OperatorProto` conversion was missing support.