Add methods to get a 1D tensor view from a lane
Having the whole lane as a 1D view is more flexible than an iterator over items
in the lane. It allows random access for example which is useful when
implementing the GatherElements and ScatterElements ONNX ops.  The ndarray
`ArrayBase::lanes` method returns an iterator where the item is a 1D view, and
that would have been a better choice here. As an interim step, add methods to
get the 1D view from the lane.