Optimize ScatterElements
Change ScatterElements to iterate over 1D lanes of data/indices/updates in
lock-step and then scatter all updates in that lane. This is much more efficient
than creating, modifying and indexing with dynamically-sized index vectors. It
also mirrors how GatherElements is implemented.

In a SageConv [^1] model, this reduced time in ScatterElements ops from
~55ms to ~12ms.

[^1]: https://huggingface.co/onnxmodelzoo/legacy_models/blob/main/Graph_Machine_Learning/sageconv_Opset16_graph_convolutions/sageconv_Opset16.onnx