Reduce number of bounds checks when indexing a tensor
Reduce the number of bounds checks when indexing a tensor from N + 1 (where N is
the number of dimensions) to N. Previously each index was checked against the
corresponding dimension size, then the offset returned by the layout was checked
against the storage length. Remove the check against the storage length, but
make it subject to the layout implementing an unsafe `TrustedLayout` trait.

Also remove `Layout::try_offset` in favor of making `offset` return an option,
which makes `offset`/`offset_unchecked` consistent with eg.
`get`/`get_unchecked` for tensors.