Support 1D-3D inputs in Resize op with up to two resized dims
When the input is not 4D and at least one dimension is resized, the input is
expanded to an NCHW tensor, resized and then added dimensions are removed. For
the special case of input_size == output_size, any rank is supported as the
input can just be copied.

For the 3D case, we differentiate between NCW (1 resized dim) and NHW (2 resized
dims) to take advantage of parallelism over channels where possible. For the 2D
case we don't differentiate between NW (or CW) and HW, but this could be added
in future.