Handle non-power-of-2 sized FFT more correctly in Whisper example
Previously the Whisper example used microfft to compute the FFT, which only
supports power-of-2 FFT sizes. Whisper however uses an FFT of size
400. This was handled by zero-padding the signal to the nearest power of two
(512). This is simple, but distorts the result a little. Switch to using
rustfft instead which can handle non-power-of-2 sizes. From brief testing,
this improves accuracy very slightly.