Implement STFT operator
Implement the STFT operator using rustfft as the FFT implementation. As this
requires additional dependencies, it is behind an `fft` crate feature flag. The
ONNX STFT operator documentation is light on details in some places so the
documentation for `torch.stft` [^1] was used as a reference to fill in gaps.

To simplify enabling all operators in `make test` and the CLI, also add an
`all-ops` feature to the rten crate.

[^1]: https://docs.pytorch.org/docs/stable/generated/torch.stft.html