Move FlatBuffers schema for model files into a separate crate
Move the FlatBuffers schema and code generated for it by the FlatBuffers
compiler into an rten-model-file crate.

The immediate motivation is to help reduce compile times for the main crate, but
logically it makes sense to separate reading a model file from computation.
Also in future it may be desirable to opt-out of support for this format once
https://github.com/robertknight/rten/issues/141 is implemented.