Correct Hann window function in Whisper example
The original implementation calls `torch.hann_window` with the `periodic`
argument implicitly set to True. This changes the denominator from `N - 1` to
`N`. Since `N = 400` here, this only has a small impact.

Also change the implementation to return a plain `Vec<f32>`. This is easier to
copy into a separate project and `NdTensor` functionality is not needed.