Add Kokoro TTS example
Add a speech-to-text example using Kokoro (https://github.com/hexgrad/kokoro).
Compared to the Piper example this is slower but produces higher quality output.

I originally used an ONNX model from Kokoro.js (see
https://huggingface.co/onnx-community/Kokoro-82M-v1.0-ONNX) but found that this
produced output that either contained NaNs or high-pitched noise. This turned
out to be due to the STFT operator producing some exactly-zero outputs in RTen's
implementation, whereas the ONNX Runtime STFT implementation produced very small
non-zero values given the same input. Adding `f32::EPSILON` to the imaginary
part of the STFT outputs in RTen worked as a hack, but is not a proper solution.

The solution for the moment is to use a different Kokoro ONNX model exported
using the scripts in the upstream repo. This replaces the STFT op with a
convolution.