Avoid storing small buffers in pool
The purpose of the buffer pool is to avoid the overhead of allocating and
freeing large buffers. For small allocations the system allocator is more
efficient. We especially want to avoid accumulating a large set of tiny buffers
because allocation uses a naive O(N) search to find matches.

In a test on macOS this made the Kokoro example about 6% faster with a typical
sentence-length input. In that case the pool got polluted with a large number of
tiny buffers, making allocations slow.

Pool allocation/hit counts reported with `RTEN_TIMING=1`:

Before: `Pool allocs 404562 hits 365927`
After: `Pool allocs 232969 hits 216970`