Use approximate algorithm in RandomNormal, RandomNormalLike
For use cases for normal distributions in ML models, fastrand-contrib's
approximate algorithm is accurate enough and much faster. Time for
RandomNormalLike ops on a test input in the Kokoro example was reduced from
~16ms to ~2.5ms.

In the test the tolerances have been reduced by increasing the sample count and
reducing the threshold.