Inline `Model::node_id` calls in examples
Reduce the boilerplate in looking up the node IDs of model inputs and outputs by
inlining `Model::node_id` calls. I did experiment with an alternative in
https://github.com/robertknight/rten/pull/933 which would allow `Model::run` to
accept names directly. However that has the downside of making the type
signature more complex and errors harder to understand if an incorrect value
passed as a node name.