Fix missing inputs for some in-place operators in timing records
If an operator supports being run in-place, but was not because the in-place
input is needed by other operators, then that input would be missing from the
profiling record. This happened because the logic to populate the `input_meta`
vec didn't properly handle the case where `in_place_input_id` was `Some(_)` but
`in_place_input` was None.

The fix is to refactor the code so that `in_place_input` is an `Option<(NodeId,
Value)>` for the actually in-placed input and downstream code always handles the
ID and value together.