Initial commit of ONNX model parsing crate
This consists of two parts:

 1. A custom low-level incremental Protocol Buffers decoder library

 2. A decoder for a subset of the `ModelProto` protobuf message type defined
    in the ONNX protobuf schemas.

The custom decoder supports incrementally reading large protobuf messages
without needing to read the whole message into memory first. This is useful as
`ModelProto` messages with embedded weights can be up to 2GB in size.
Incremental decoding enables reading weights from the file directly into aligned
buffers without additional copies. The decoder also supports skipping over
fields, which is used to avoid reading fields that are not necessary for
inference.

For ONNX models using external weights, the custom decoder provides less of a
benefit, as the protobuf message has a "normal" size in that case and the
external weights have an alignment that enables them to be memory-mapped.