Add initial implementation of ONNX model loader
Add an initial implementation of an ONNX model loader. This supports all
operators and data types that are currently supported in RTen. The decision of
which loader to use is based on the file extension. For backwards compatibility
the file type is assumed to be `.rten` in `Model::load` where no file type is
specified.

Some limitations of this initial implementation:

 1. Tensors stored in external data files are not supported.
 2. Unused operator attributes are not detected as they are by the
    rten-convert tool
 3. f16 tensors are not supported
 4. The error messages reported when protobuf parsing fails are poor
 5. Loading ONNX models from an in-memory buffer is not yet supported
 6. Promotion of attributes to inputs is not implemented yet. This
    affects older models with Clip, Split, Squeeze and Unsqueeze
    operators. See uses of `generate_input_from_attr` in rten-convert.
 7. No automated tests yet. I manually tested various examples swapping
    out .rten models for .onnx models.