Use shared ownership for more constants
The preferred backing storage for constant data is either an mmap-ed buffer or a
slice of a `Vec<u8>` into which the model file was originally loaded, as these
options minimize copying when loading a model. When loading weights from a
`.rten` file, weights are usually stored this way.

In cases where these options are not available, unique ownersip is currently
used. However this prevents sharing of constants between sub-graphs, which in
turn limits constant propagation in sub-graphs. To remedy this, allow constants
to use an `ArcTensor<T>` as the backing storage and change existing creation of
owned constants to use this where possible.

One place where unique ownership cannot be replaced with shared ownership
trivially is after constant propagation. `Graph::partial_run` returns a set of
owned tensors which we want to turn into constants. Turning these into
reference-counted constants would occur an additional copy.