Fix panic in ONNX loader if graph input or output has empty name
Return an error instead of panicking if a graph input or output has an empty
name. The panic occurred because the loop which creates values skipped over
values with empty names, but we later assumed that nodes would exist in the RTen
graph for all inputs and outputs in the ONNX graph.