Fix loading of zero-length tensors in the ONNX loader
When a tensor has zero length, the `TensorProto.raw_data` field can be an empty
`Vec<u8>` whose data pointer is a dangling pointer with a smaller alignment that
the global allocator's minimum, causing `ArcSlice::from_bytes` to return `None`.
Address this in `ArcSlice::from_bytes` by creating a dummy Vec with zero length
but non-zero capacity which will allocate. The same issue could in principle
also happen for tensors with external data when using `FileLoader`. Handle that
be delegating to `ArcSlice::from_bytes` in `tensor_from_external_data`.