Update library documentation to mention ONNX format support
Update the front page of the rten crate docs to mention and recommend support
for loading models in ONNX format directly, rather than converting to `.rten`
first. This simplifies getting started with the library. Converting to `.rten`
format is an optimization that can be applied later.

Also update the "Preparing models" section with updated advice on where to get
ONNX models from.