Convert Nougat example to use ONNX format
For consistency with other examples, convert the Nougat example to load the
model in ONNX format. These models are large in fp32 (1GB for the encoder), so
this is a case where converting to .rten format would be beneficial. This has
been noted in a comment. For ONNX files not using external data, using
`Model::load_file` is slightly faster.