Add model format compatibility notes for `Model::load_mmap`
Note that `load_mmap` doesn't offer a benefit for `.onnx` files with embedded
weights, since mmap-ing an ONNX file results in the tensor data in
`TensorProto.raw_data` not having sufficient alignment. Hence the weights have
to be copied into buffers in private memory anyway. The solution is either to
convert the model to use external data, or convert to .rten format.