Check for overflow on narrowing conversions in ONNX loader
Use checked casts for most narrowing conversions in the ONNX loader. This will
make for easier debugging if an attribute expected to have a positive value has
a negative value for example.

Disallow `as` casts in this module as a cue for code to be careful about whether
casts should error, saturate or wrap.