Check for unused attributes in ONNX loader
Make ONNX model loading fail if an operator has attributes which are not read,
as this indicates unsupported functionality. This mirrors the behavior in
rten-convert which warns about unsupported attributes when converting models.
This is currently a hard error, although in future it might be useful to have a
way to ignore certain attributes.

Add logic to ignore default values for various attributes when they are set
explicitly, mirroring uses of `attr_reader.check_attr` in rten-convert.