Fix model loading when value info is present for captured value
Fix models failing to load if a subgraph contains value info in
`GraphProto.value_info` for a value that is captured from the parent graph. When
iterating over `value_info` a value node was created for each entry. This entry
caused the logic for marking nodes as captures to be skipped when iterating over
operator inputs and outputs later. As a result the value was incorrectly not
marked as a capture and planning would fail because the value was not available.

The fix is to build a name -> metadata map when traversing `value_info` but only
actually create value nodes when a value is encountered as an operator
input/output, at which point we check if it is a capture.