Revise ModelMetadata to be model-format independent
The rten model format has a set of standard fields, whereas ONNX uses free-form
key/value pairs. Change the implementation of the ModelMetadata type so that it
can support both of these. The public API retains methods for individual fields
for backwards compatibility, plus there is a new API for iterating over fields
as a collection.

The display of field names in the CLI has changed to use identifier-like names
("code_repository" instead of "Code repository"). This will make the
presentation of standard and custom fields more similar.