Revise error type for model load errors
Replace the `ModelLoadError` enum with a `LoadError` struct which has a design
more like `std::io::Error`. The goal is to decouple the internal structure from
the public API to make it easier to avoid breaking changes in future.
`LoadError` has a `ModelLoadError` alias which will make this a non-breaking
change for many downstream consumers.