Revise error type for model inference errors
Revise the RunError type along the same lines as the LoadError changes in
https://github.com/robertknight/rten/pull/1007. Change `RunError` from an enum
to a struct with properties for the error category and the node path.

These changes separate the implementation and public API of RunError to add
flexibility to add new errors or error details in future without a breaking API
change.