Revise API for enabling operator subsets
Revise the API for loading a model with only a subset of operators enabled. The
new API is easier to use and decouples the types/traits used to enable an
operator from those that implement deserialization and evaluation internally.
This is part of a general effort to expose less of rten's internals in the
public API.

Usage of the new API looks like so:

```
use rten::{ModelOptions, op_registry};

let registry = op_registry!(Conv, Add, Relu, MatMul);
let model = ModelOptions::with_ops(registry).load_file(path)?;
```

A result of this change is that the `ReadOpError` type is no longer exposed in
the crate's public API.