Remove model_builder crate feature and associated public API
Downstream usage of the model builder API (that I know about) was removed in
https://github.com/robertknight/ocrs/pull/206. The API is still used internally
for tests but no longer needs to be exposed.

An upside of this change is that the linter can now identify that some
`rten_builder::OpType` variants are not currently used. These have been marked
as `#[allow(dead_code)]` for now as a reminder.