Rename `CastError` to `TryFromValueError` and export it
The `CastError` error type used when converting `Value`/`ValueView`s to tensors
is encountered by users but was not exported in the public API so consumers were
limited in how they could use it. Rename to `TryFromValueError` to align with
similar types in std (eg. `TryFromIntError`) and export it. Also improve the
description of each error variant and mark as non_exhaustive to enable adding
more errors in future (eg. for when ONNX optional or map types are added).