Simplify model input/output creation/extraction
Add methods to create `Value`s and `ValueView`s directly from shape and data, as
well as to extract the shape and data from a value. This enables projects to use
rten without a direct dependency on rten-tensor. Also improve the documentation
around creating values and extracting the shape/data from them.