Make `Value::dtype` return collection and element type
Change `Value::dtype` and `ValueView::dtype` to return a type which includes
both the collection type (tensor or sequence) as well as the element type of
tensor elements (f32, i32 etc.).

Adjust consumers of this method to handle the new cases appropriately. As a
result the CastLike, EyeLike and SequenceInsert operators will now error if they
get a sequence input where a tensor is expected. This aligns better with the
ONNX specifications for these operators.